Easily Query nycflights13-Like Air Travel Data

The anyflights function is a wrapper around the following functions:

get_airlines: Grab data to translate between two letter carrier codes and names

get_airports: Grab data on airport names and locations

get_flights: Grab data on all flights that departed given airports in a given year and month

get_planes: Grab construction information about each plane

get_weather: Grab hourly meterological data for a given airport in a given year and month

anyflights(station, year, month = 1:12, dir = NULL)

get_airlines(dir = NULL, flights_data = NULL)

get_airports(dir = NULL)

get_flights(station, year, month = 1:12, dir = NULL)

get_planes(year, dir = NULL, flights_data = NULL)

get_weather(station, year, month, dir = NULL)

Arguments

station
A character vector giving the origin airports of interest (as the FAA LID airport code).

year
A numeric giving the year of interest. This argument is currently not vectorized, as dataset sizes for single years are significantly large. Information for the most recent year is usually available by February or March in the following year.

month
A numeric giving the month(s) of interest.

dir
An (always) optional character string giving the directory to save datasets in. By default, datasets will not be saved to file.

flights_data
An optional argument to get_airlines and get_planesso that the function will only return carriers that appear in the flight data of interest---either a filepath as a character string or a dataframe outputted by get_flights. If not supplied, all carriers and planes will be returned.

Value

A named list of dataframes and, optionally, a folder containing the datasets saved to file.

Details

The function returns a list of dataframes (or directory of datasets) similar to those found in the nycflights13 data package. Please note that, even with a strong internet connection, this function may take several minutes to download relevant data.

Examples