Generate a Data Package from `anyflights` Data

Generate a data-only package, including documentation, from data outputted by the `anyflights()` function. Please do not submit the outputted package to CRAN or similar repositories as original packages.

Usage

as_flights_package(data, name = make.names(deparse(substitute(data))))

Arguments

data

A named list of dataframes outputted by anyflights.

name

The desired name of the resulting package as a character string. The package will check that the supplied package name is valid using the regular expression .standard_regexps()$valid_package_name, and save the output in a directory by the same name. Defaults to make.names(deparse(substitute(data))).

Value

A directory containing a data-only package built around the supplied data.